Refactor date utils
