+++ src/shared/date/core.ts
@@ -1,0 +1,5 @@
+export function addDays(date:Date, numDays:number) {
+    const newDate = new Date(date);
+    newDate.setDate(newDate.getDate() + numDays);
+    return newDate;
+}
