+++ src/shared/date/equate.ts
@@ -1,0 +1,5 @@
+export function equalDate(a: Date, b: Date) {
+    return a.getFullYear() === b.getFullYear() &&
+        a.getMonth() === b.getMonth() &&
+        a.getDate() === b.getDate();
+}
