+++ src/shared/date/format.ts
@@ -1,0 +1,3 @@
+export function toISODate(date: Date): string {
+    return date.toISOString().split('T')[0];
+}
